requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
ecdsa==0.19.0
pillow==10.3.0
pyasn1==0.6.0
PyJWT==2.8.0
pypng==0.20220715.0
python-jose==3.3.0
qrcode==7.4.2
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
